Struct rusoto_sqs::SendMessageResult

pub struct SendMessageResult {
    pub md5_of_message_attributes: Option<String>,
    pub md5_of_message_body: Option<String>,
    pub message_id: Option<String>,
    pub sequence_number: Option<String>,
}

The MD5OfMessageBody and MessageId elements.

Fields
md5_of_message_attributes: Option<String>

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

md5_of_message_body: Option<String>

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

message_id: Option<String>

An attribute containing the MessageId of the message sent to the queue. For more information, see Queue and Message Identifiers in the Amazon Simple Queue Service Developer Guide.

sequence_number: Option<String>

This parameter applies only to FIFO (first-in-first-out) queues.

The large, non-consecutive number that Amazon SQS assigns to each message.

The length of SequenceNumber is 128 bits. SequenceNumber continues to increase for a particular MessageGroupId.