Struct rusoto_sqs::SqsClient

pub struct SqsClient { /* fields omitted */ }

A client for the Amazon SQS API.

Methods

impl SqsClient

pub fn new(region: Region) -> SqsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SqsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,