Crate rusoto_sqs

Welcome to the Amazon Simple Queue Service API Reference.

Amazon Simple Queue Service (Amazon SQS) is a reliable, highly-scalable hosted queue for storing messages as they travel between applications or microservices. Amazon SQS moves data between distributed application components and helps you decouple these components.

Standard queues are available in all regions. FIFO queues are available in the US East (N. Virginia), US East (Ohio), US West (Oregon), and EU (Ireland) regions.

You can use AWS SDKs to access Amazon SQS using your favorite programming language. The SDKs perform tasks such as the following automatically:

Cryptographically sign your service requests

Retry requests

Handle error responses

Additional Information

Amazon SQS Product Page

Amazon Simple Queue Service Developer Guide

Making API Requests

Using Amazon SQS Message Attributes

Using Amazon SQS Dead-Letter Queues

Amazon Web Services General Reference

Regions and Endpoints

If you're using the service, you're probably looking for SqsClient and Sqs.

Structs

AddPermissionRequest

BatchResultErrorEntry
This is used in the responses of batch API to give a detailed description of the result of an action on each entry in the request.

ChangeMessageVisibilityBatchRequest

ChangeMessageVisibilityBatchRequestEntry
Encloses a receipt handle and an entry id for each message in ChangeMessageVisibilityBatch.

All of the following list parameters must be prefixed with ChangeMessageVisibilityBatchRequestEntry.n, where n is an integer value starting with 1. For example, a parameter list for this action might look like this:

&amp;ChangeMessageVisibilityBatchRequestEntry.1.Id=change_visibility_msg_2

&amp;ChangeMessageVisibilityBatchRequestEntry.1.ReceiptHandle=<replaceable>Your_Receipt_Handle</replaceable>

&amp;ChangeMessageVisibilityBatchRequestEntry.1.VisibilityTimeout=45

ChangeMessageVisibilityBatchResult
For each message in the batch, the response contains a ChangeMessageVisibilityBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

ChangeMessageVisibilityBatchResultEntry
Encloses the Id of an entry in ChangeMessageVisibilityBatch.

ChangeMessageVisibilityRequest

CreateQueueRequest

CreateQueueResult
Returns the QueueUrl attribute of the created queue.

DeleteMessageBatchRequest

DeleteMessageBatchRequestEntry
Encloses a receipt handle and an identifier for it.

DeleteMessageBatchResult
For each message in the batch, the response contains a DeleteMessageBatchResultEntry tag if the message is deleted or a BatchResultErrorEntry tag if the message can't be deleted.

DeleteMessageBatchResultEntry
Encloses the Id of an entry in DeleteMessageBatch.

DeleteMessageRequest

DeleteQueueRequest

GetQueueAttributesRequest

GetQueueAttributesResult
A list of returned queue attributes.

GetQueueUrlRequest

GetQueueUrlResult
For more information, see Responses in the Amazon Simple Queue Service Developer Guide.

ListDeadLetterSourceQueuesRequest

ListDeadLetterSourceQueuesResult
A list of your dead letter source queues.

ListQueueTagsRequest

ListQueueTagsResult

ListQueuesRequest

ListQueuesResult
A list of your queues.

Message
An Amazon SQS message.

MessageAttributeValue
The user-specified message attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see SendMessage.

Name, type, value and the message body must not be empty or null. All parts of the message attribute, including Name, Type, and Value, are part of the message size restriction (256 KB or 262,144 bytes).

PurgeQueueRequest

ReceiveMessageRequest

ReceiveMessageResult
A list of received messages.

RemovePermissionRequest

SendMessageBatchRequest

SendMessageBatchRequestEntry
Contains the details of a single Amazon SQS message along with an Id.

SendMessageBatchResult
For each message in the batch, the response contains a SendMessageBatchResultEntry tag if the message succeeds or a BatchResultErrorEntry tag if the message fails.

SendMessageBatchResultEntry
Encloses a MessageId for a successfully-enqueued message in a SendMessageBatch.

SendMessageRequest

SendMessageResult
The MD5OfMessageBody and MessageId elements.

SetQueueAttributesRequest

SqsClient
A client for the Amazon SQS API.

TagQueueRequest

UntagQueueRequest

Enums

AddPermissionError
Errors returned by AddPermission

ChangeMessageVisibilityBatchError
Errors returned by ChangeMessageVisibilityBatch

ChangeMessageVisibilityError
Errors returned by ChangeMessageVisibility

CreateQueueError
Errors returned by CreateQueue

DeleteMessageBatchError
Errors returned by DeleteMessageBatch

DeleteMessageError
Errors returned by DeleteMessage

DeleteQueueError
Errors returned by DeleteQueue

GetQueueAttributesError
Errors returned by GetQueueAttributes

GetQueueUrlError
Errors returned by GetQueueUrl

ListDeadLetterSourceQueuesError
Errors returned by ListDeadLetterSourceQueues

ListQueueTagsError
Errors returned by ListQueueTags

ListQueuesError
Errors returned by ListQueues

PurgeQueueError
Errors returned by PurgeQueue

ReceiveMessageError
Errors returned by ReceiveMessage

RemovePermissionError
Errors returned by RemovePermission

SendMessageBatchError
Errors returned by SendMessageBatch

SendMessageError
Errors returned by SendMessage

SetQueueAttributesError
Errors returned by SetQueueAttributes

TagQueueError
Errors returned by TagQueue

UntagQueueError
Errors returned by UntagQueue

Traits

Sqs
Trait representing the capabilities of the Amazon SQS API. Amazon SQS clients implement this trait.